updated layout for new sidebar style
