+++ src/app/(protected)/(sidebar)/layout.tsx
@@ -1,0 +1,28 @@
+"use client"
+
+import { AppSidebar } from "@/components/app-sidebar"
+import {
+  SidebarInset,
+  SidebarProvider,
+} from "@/components/ui/sidebar"
+
+export default function SidebarLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return (
+    <main className="min-h-screen bg-transparent relative">
+      <SidebarProvider>
+        <AppSidebar />
+        <SidebarInset>
+          {children}
+        </SidebarInset>
+      </SidebarProvider>
+      
+      {/* Background texture */}
+      <div
+        className="fixed inset-0 -z-10 bg-[radial-gradient(#e5e7eb_1px,transparent_1px)] bg-[length:20px_20px] opacity-50"></div>
+    </main>
+  )
+}
