+++ src/app/(protected)/(sidebar)/report/[id]/page.tsx
