--- src/app/(protected)/report/layout.tsx
@@ -1,7 +1,0 @@
-export default function ReportLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return children
-}
